requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
clearml==1.16.5
clearml-agent==1.16.5
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
Crawl4AI==0.3.742
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
executing==2.1.0
fake-http-header==0.3.5
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
furl==2.1.3
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.0.3
grpcio==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
litellm==1.52.14
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mockito==1.5.3
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
openai==1.55.0
orderedmultidict==1.0.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib2==2.3.7.post1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
playwright==1.47.0
pluggy==1.5.0
portalocker==2.10.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.1
pydantic_core==2.27.1
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.10.1
pyparsing==3.2.0
pytest==8.3.3
pytest-mockito==0.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.1.1
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
ruff==0.8.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==2.2.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tf-playwright-stealth==1.1.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.1.0
tomlkit==0.12.0
torch==2.5.1
tornado==6.4.2
tqdm==4.65.0
traitlets==5.14.3
transformers==4.46.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
websockets==12.0
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0
